Split C++ errors from webview.h
