+++ core/include/webview/detail/backends/gtk_webkitgtk.hh
@@ -38,6 +38,9 @@
 //
 // ====================================================================
 //
+
+#include "../../errors.hh"
+
 #include <cstdlib>
 
 #include <gtk/gtk.h>
