+++ core/include/webview/webview.h
@@ -36,6 +36,7 @@
 #ifdef __cplusplus
 #ifndef WEBVIEW_HEADER
 
+#include "errors.hh"
 #include "detail/basic_result.hh"
 
 #include <algorithm>
@@ -68,50 +69,6 @@
 namespace webview {
 
 using dispatch_fn_t = std::function<void()>;
-
-class error_info {
-public:
-  error_info(webview_error_t code, const std::string &message = {}) noexcept
-      : m_code{code}, m_message{message} {}
-  error_info() = default;
-
-  webview_error_t code() const { return m_code; }
-  const std::string &message() const { return m_message; }
-
-private:
-  webview_error_t m_code{WEBVIEW_ERROR_UNSPECIFIED};
-  std::string m_message;
-};
-
-class exception : public std::exception {
-public:
-  exception(webview_error_t code, const std::string &message,
-            std::exception_ptr cause) noexcept
-      : exception{error_info{code, message}, cause} {}
-
-  exception(webview_error_t code, const std::string &message) noexcept
-      : exception{error_info{code, message}} {}
-
-  exception(const error_info &error, std::exception_ptr cause) noexcept
-      : m_error{error},
-        // NOLINTNEXTLINE(bugprone-throw-keyword-missing)
-        m_cause{cause} {}
-
-  exception(const error_info &error) noexcept : m_error{error} {}
-
-  exception() = default;
-
-  const error_info &error() const { return m_error; }
-  std::exception_ptr cause() const { return m_cause; }
-
-  const char *what() const noexcept override {
-    return m_error.message().c_str();
-  }
-
-private:
-  error_info m_error{WEBVIEW_ERROR_UNSPECIFIED};
-  std::exception_ptr m_cause;
-};
 
 template <typename T>
 using result = detail::basic_result<T, error_info, exception>;
